+++ src/main/java/de/p2tools/mtplayer/controller/starter/RuntimeExec.java
@@ -230,7 +230,7 @@ public class RuntimeExec {
                     final int diffTime = download.getStart().getStartTime().diffInSeconds();
                     final double diffPercent = percent - percent_start;
                     final double restPercent = DownloadConstants.PROGRESS_FINISHED - percent;
-                    download.getStart().setTimeLeftSeconds((long) (diffTime * restPercent / diffPercent));
+                    download.getStart().setTimeLeftSeconds((int) (diffTime * restPercent / diffPercent));
                 }
             }
         }
